Fix #8 by removing postgres port map
We don’t need to expose the postgres port on the host. The containers communicate via their own ports on the bridge network. Exposing the postgres port on `5432` can lead to port clashes with existing postgres instances. If we want to expose postgres, we should do it on something like port `54321`, *but* we don’t need to change the connect strings etc elsewhere as they still communicate via `5432` on the bridge network.